backend/requirements.txt
Django==5.1
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.4.0
psycopg[binary]==3.2.4
python-dotenv==1.0.0
google-generativeai==0.8.4
numpy==2.2.3
requests==2.32.3
